Clean up stuff for release on wp.org
