+++ .distignore
@@ -1,0 +1,92 @@
+# Development and testing files
+/.git
+/.gitignore
+/.gitmodules
+/.distignore
+/tests
+/test
+/phpunit.xml
+/phpunit.xml.dist
+/phpcs.xml
+/phpstan.neon
+/.travis.yml
+/.github
+/bin
+/Gruntfile.js
+/gulpfile.js
+/package.json
+/package-lock.json
+/composer.json
+/composer.lock
+/webpack.config.js
+/tsconfig.json
+
+# Node modules and build tools
+/node_modules
+/bower_components
+
+# Vendor development files
+/vendor/*/tests
+/vendor/*/test
+/vendor/*/Tests
+/vendor/*/Test
+/vendor/*/.git
+/vendor/*/.github
+/vendor/*/docs
+/vendor/*/doc
+/vendor/*/Docs
+/vendor/*/Doc
+/vendor/*/examples
+/vendor/*/Examples
+/vendor/*/example
+/vendor/*/Example
+/vendor/*/demo
+/vendor/*/Demo
+/vendor/*/demos
+/vendor/*/Demos
+/vendor/*/.travis.yml
+/vendor/*/.travis-ci-apache
+/vendor/*/Gemfile
+/vendor/*/Gemfile.lock
+/vendor/*/LICENSE
+/vendor/*/license
+/vendor/*/README
+/vendor/*/readme
+/vendor/*/README.md
+/vendor/*/readme.md
+/vendor/*/CHANGELOG
+/vendor/*/changelog
+/vendor/*/CHANGELOG.md
+/vendor/*/changelog.md
+/vendor/*/CONTRIBUTING
+/vendor/*/contributing
+/vendor/*/CONTRIBUTING.md
+/vendor/*/contributing.md
+/vendor/*/phpunit.xml
+/vendor/*/phpunit.xml.dist
+/vendor/*/phpcs.xml
+/vendor/*/composer.json
+/vendor/*/composer.lock
+/vendor/*/package.json
+/vendor/*/package-lock.json
+/vendor/*/*.h
+/vendor/*/API_VERSION
+
+# Specific problematic vendor files mentioned in review
+/vendor/amphp/sync/docs/Gemfile
+/vendor/amphp/dns/docs/Gemfile
+/vendor/stripe/stripe-php/API_VERSION
+/vendor/amphp/hpack/src/Internal/amp-hpack.h
+/vendor/amphp/socket/docs/Gemfile
+/vendor/delight-im/cookie/.travis-ci-apache
+/vendor/doctrine/deprecations/src/PHPUnit
+
+# Other development files
+*.log
+*.tmp
+*.bak
+*.swp
+*.swo
+*~
+.DS_Store
+Thumbs.db
