+++ assets/css/site-maintenance.css
@@ -1,0 +1,4 @@
+.wu-styling h3 {
+	font-weight: 600 !important;
+	font-size: 90% !important;
+}
