+++ assets/css/tax-details.css
@@ -1,0 +1,4 @@
+#wu-line_item_list_table .tablenav.bottom,
+#wu-line_item_list_table tfoot {
+	display: none;
+}
