+++ assets/css/toolbox.css
@@ -1,0 +1,12 @@
+#wu-toolbox-links {
+	transition: width 2s;
+}
+.wu-toolbox-closed #wu-toolbox-links {
+	transition: width 2s;
+	width: 1px;
+	height: 1px;
+	overflow: hidden;
+}
+#wu-toolbox-membership-status {
+	margin-top: 2px;
+}
