--- inc/admin-pages/class-about-admin-page.php
@@ -1,147 +1,0 @@
-<?php
-/**
- * Multisite Ultimate About Admin Page.
- *
- * @package WP_Ultimo
- * @subpackage Admin_Pages
- * @since 2.0.0
- */
-
-namespace WP_Ultimo\Admin_Pages;
-
-// Exit if accessed directly
-defined('ABSPATH') || exit;
-
-/**
- * Multisite Ultimate About Admin Page.
- */
-class About_Admin_Page extends Base_Admin_Page {
-
-	/**
-	 * Holds the ID for this page, this is also used as the page slug.
-	 *
-	 * @var string
-	 */
-	protected $id = 'wp-ultimo-about';
-
-	/**
-	 * Is this a top-level menu or a submenu?
-	 *
-	 * @since 1.8.2
-	 * @var string
-	 */
-	protected $type = 'submenu';
-
-	/**
-	 * Is this a top-level menu or a submenu?
-	 *
-	 * @since 1.8.2
-	 * @var string
-	 */
-	protected $parent = 'none';
-
-	/**
-	 * This page has no parent, so we need to highlight another sub-menu.
-	 *
-	 * @since 2.0.0
-	 * @var string
-	 */
-	protected $highlight_menu_slug = 'wp-ultimo';
-
-	/**
-	 * Should we hide admin notices on this page?
-	 *
-	 * @since 2.0.0
-	 * @var boolean
-	 */
-	protected $hide_admin_notices = true;
-
-	/**
-	 * Should we force the admin menu into a folded state?
-	 *
-	 * @since 2.0.0
-	 * @var boolean
-	 */
-	protected $fold_menu = true;
-
-	/**
-	 * Holds the admin panels where this page should be displayed, as well as which capability to require.
-	 *
-	 * To add a page to the regular admin (wp-admin/), use: 'admin_menu' => 'capability_here'
-	 * To add a page to the network admin (wp-admin/network), use: 'network_admin_menu' => 'capability_here'
-	 * To add a page to the user (wp-admin/user) admin, use: 'user_admin_menu' => 'capability_here'
-	 *
-	 * @since 2.0.0
-	 * @var array
-	 */
-	protected $supported_panels = [
-		'network_admin_menu' => 'manage_network',
-	];
-
-	/**
-	 * Returns the title of the page.
-	 *
-	 * @since 2.0.0
-	 * @return string Title of the page.
-	 */
-	public function get_title() {
-
-		return __('About', 'multisite-ultimate');
-	}
-
-	/**
-	 * Returns the title of menu for this page.
-	 *
-	 * @since 2.0.0
-	 * @return string Menu label of the page.
-	 */
-	public function get_menu_title() {
-
-		return __('Multisite WaaS', 'multisite-ultimate');
-	}
-
-	/**
-	 * Allows admins to rename the sub-menu (first item) for a top-level page.
-	 *
-	 * @since 2.0.0
-	 * @return string False to use the title menu or string with sub-menu title.
-	 */
-	public function get_submenu_title() {
-
-		return __('About', 'multisite-ultimate');
-	}
-
-	/**
-	 * Every child class should implement the output method to display the contents of the page.
-	 *
-	 * @since 1.8.2
-	 * @return void
-	 */
-	public function output(): void {
-
-		wu_get_template('about');
-	}
-
-	/**
-	 * Adds the cure bg image here as well.
-	 *
-	 * @since 2.0.0
-	 * @return void
-	 */
-	public function register_scripts(): void {
-
-		parent::register_scripts();
-
-		wp_add_inline_style(
-			'wu-admin',
-			sprintf(
-				'
-		#wpwrap {
-			background: url("%s") right bottom no-repeat;
-			background-size: 60%%;	 
-		}',
-				wu_get_asset('bg-setup.webp', 'img')
-			)
-		);
-	}
-}
