+++ inc/class-light-ajax.php
@@ -101,7 +101,8 @@ class Light_Ajax {
 			]
 		);
 
-		$action = isset($_REQUEST['wu-when']) ? base64_decode((string) $_REQUEST['wu-when']) : 'plugins_loaded'; // phpcs:ignore WordPress
+		// nonce checked in process_light_ajax().
+		$action = isset($_REQUEST['wu-when']) ? base64_decode(sanitize_key(wp_unslash($_REQUEST['wu-when']))) : 'plugins_loaded'; // phpcs:ignore WordPress
 
 		return in_array($action, $allowed_list, true) ? $action : 'plugins_loaded';
 	}
