+++ views/dashboard-widgets/site-maintenance.php
@@ -16,9 +16,3 @@ defined( 'ABSPATH' ) || exit;
 
 </div>
 
-<style>
-.wu-styling h3 {
-	font-weight: 600 !important;
-	font-size: 90% !important;
-}
-</style>
