+++ views/settings/fields/field-color.php
@@ -42,11 +42,7 @@ defined( 'ABSPATH' ) || exit;
 
 </div>
 
-<script type="text/javascript">
-(function($) {
-	$(function() {
-		// Add Color Picker to all inputs that have 'color-field' class
-		$('.field-<?php echo esc_attr($field->id); ?>').wpColorPicker();
-	});
-})(jQuery);
-</script>
+<?php
+wp_enqueue_script('wu-color-field', wu_get_asset('color-field.js', 'js'), ['jquery', 'wp-color-picker'], wu_get_version(), true);
+wp_add_inline_script('wu-color-field', 'var wu_color_field_ids = wu_color_field_ids || []; wu_color_field_ids.push("' . esc_js($field->id) . '");', 'before');
+?>
